refactor: new configure way
