+++ lua/ltex_extra/test/init.lua
@@ -1,0 +1,5 @@
+require("ltex_extra").setup()
+P(package.loaded.ltex_extra)
+package.loaded.ltex_extra = nil
+
+
